fix: resolve lint errors from CI — unused import and non-component export
Remove unused DashboardFilterSnapshot import from DashboardShell.
Move TAB_LABELS to dashboard/tabLabels.ts so the component file only
exports components (satisfies react-refresh/only-export-components).
