+++ src/components/dashboard/tabLabels.ts
@@ -1,0 +1,14 @@
+import type { Tab } from './types';
+
+export const TAB_LABELS: Record<Tab, string> = {
+  telemetry: 'Telemetry',
+  tires:     'Tyres & Strategy',
+  energy:    'DRS, Gears & RPM',
+  trackmap:  'Track Map',
+  positions: 'Race Positions',
+  intervals: 'Intervals',
+  radio:     'Team Radio',
+  incidents: 'Race Control',
+  weather:   'Weather',
+  broadcast: 'Broadcast',
+};
